enum_dispatch for payload storage
