+++ lib/segment/src/payload_storage/mod.rs
@@ -3,5 +3,7 @@ pub mod in_memory_payload_storage;
 mod payload_storage_base;
 pub mod query_checker;
 pub mod simple_payload_storage;
+pub mod simple_payload_storage_impl;
+pub mod in_memory_payload_storage_impl;
 
 pub use payload_storage_base::*;
